requirements.txt
fastapi
uvicorn
firebase-admin
streamlit
requests
python-dotenv
pydantic